--- src/components/UserIcon/UserIcon.module.css
@@ -1,30 +1,0 @@
-.userIcon {
-  --userIcon-size: 2rem;
-  min-width: var(--userIcon-size);
-  min-height: var(--userIcon-size);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  border-radius: 50%;
-  border: #23262a 2px solid;
-  background-color: #00315d;
-  color: #ffffff;
-  text-align: center;
-  text-transform: uppercase;
-}
-
-.lg {
-  --userIcon-size: 2.5rem;
-}
-
-.md {
-  --userIcon-size: 2rem;
-}
-
-.sm {
-  --userIcon-size: 1.75rem;
-}
-
-.icon {
-  color: #ffffff;
-}
