+++ src/features/amUI/userRightsPage/DelegationModal/DelegationModal.module.css
@@ -1,0 +1,4 @@
+.modalDialog {
+  max-width: 1100px;
+  padding: 0 15px 20px 15px;
+}
