+++ src/features/amUI/userRightsPage/SingleRightsSection.module.css
@@ -1,7 +1,7 @@
 .singleRightsSectionContainer {
   display: flex;
   flex-direction: column;
-  gap: 1rem;
+  gap: 0.2rem;
   padding: 2rem 0 0.5rem 0;
 }
 
@@ -17,7 +17,7 @@
   grid-template-areas:
     'icon title type action'
     'icon owner type action';
-  grid-template-columns: 30px minmax(30px, auto) minmax(30px, 120px) 30px;
+  grid-template-columns: 35px minmax(30px, auto) minmax(30px, 120px) 30px;
   gap: 0.5rem;
   align-items: center;
   padding: 0.5rem 0;
@@ -55,11 +55,16 @@
 }
 
 .singleRightItem > .icon {
-  font-size: 2.1rem;
   grid-area: icon;
+}
+
+.tools {
+  display: flex;
+  justify-content: space-between;
 }
 
 .pagination {
   display: flex;
   justify-content: center;
+  align-items: center;
 }
