+++ src/localizations/no_nb.json
@@ -208,6 +208,11 @@
     "single_rights_title": "Tjenester",
     "resource_type_text": "Enkelttjeneste"
   },
+  "delegation_modal": {
+    "give_to_name": "Gi <strong>{{name}}</strong> fullmakt til ny tjeneste",
+    "name_will_receive": "<strong>{{name}}</strong> vil få:",
+    "give_poa": "Gi fullmakt"
+  },
   "user_role": {
     "SREVA": "Revisor registrert i revisorregisteret",
     "FGRP": "Inngår i foretaksgruppe med",
