+++ src/localizations/no_nn.json
@@ -207,6 +207,11 @@
     "single_rights_title": "Tjenester",
     "resource_type_text": "Enkelttjeneste"
   },
+  "delegation_modal": {
+    "give_to_name": "Gi <strong>{{name}}</strong> fullmakt til ny teneste",
+    "name_will_receive": "<strong>{{name}}</strong> vil få:",
+    "give_poa": "Gi fullmakt"
+  },
   "user_role": {
     "SREVA": "Revisor registrert i revisorregisteret",
     "FGRP": "Inngår i føretaksgruppe med",
